Authorization

+++ app/Http/Controllers/ArticleCategoryController.php
@@ -4,7 +4,9 @@
 namespace App\Http\Controllers;
 use App\Interfaces\ArticleCategoryRepositoryInterface;
 use Illuminate\Http\Request;
-
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use App\Repositories\ArticleRepository;
 
 class ArticleCategoryController extends Controller
 {
@@ -32,6 +34,12 @@ class ArticleCategoryController extends Controller
 
     public function store($articleId, $categoryId)
     {
+        $articleRepository = new ArticleRepository();
+
+        $article = $articleRepository->find($articleId);
+
+        if(Auth::user()->id !== $article->user_id)
+            return response(['message' => 'Unauthorized'], 401);
 
         $articleCategory = $this->articleCategoryRepository->storeArticleCategory($articleId, $categoryId);
 
@@ -41,6 +49,13 @@ class ArticleCategoryController extends Controller
 
     public function destroy($articleId, $categoryId)
     {
+        $articleRepository = new ArticleRepository();
+
+        $article = $articleRepository->find($articleId);
+
+        if(Gate::denies('delete', $article))
+            return response(['message' => 'Unauthorized'], 401);
+
         return $this->articleCategoryRepository->deleteArticleCategory($articleId, $categoryId) ?
             response(['message' => 'Article Category deleted successfully'], 201) :
             response(['message' => 'Article or Category not found'], 404);
