+++ app/Http/Controllers/ArticleController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Interfaces\ArticleRepositoryInterface;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller {
 
@@ -38,45 +40,46 @@ class ArticleController extends Controller {
 
     public function store(Request $request)
     {
+        $user = Auth::user();
         $validatedRequest = $this->validator($request);
 
-        return $this->articlesRepository->storeArticle($validatedRequest) ?
+        return $this->articlesRepository->storeArticle($validatedRequest, $user->id) ?
             response(['message' => 'Article added successfully'], 201) :
             response(['message' => 'User not found'], 404);
     }
 
     public function update(Request $request, $articleId)
     {
+        $article = $this->articlesRepository->find($articleId);
+        if(Gate::denies('update', $article))
+            return response(['message' => 'Unauthorized'], 401);
+
         $validatedRequest = $this->validator($request);
 
         return $this->articlesRepository->updateArticle($validatedRequest, $articleId) ?
             response(['message' => 'Article updated successfully'], 201):
-            response(['message' => 'Article not found'], 404);;
+            response(['message' => 'Article not found'], 404);
     }
 
     public function destroy($articleId)
     {
+        $article = $this->articlesRepository->find($articleId);
+
+        if(Gate::denies('delete', $article))
+            return response(['message' => 'Unauthorized'], 401);
+
         return $this->articlesRepository->deleteArticle($articleId) ?
             response(['message' => 'Article deleted successfully'], 201):
             response(['message' => 'Article not found'], 404);
 
     }
 
-
-
     public function validator(Request $request)
     {
         return $this->validate($request, [
-            'user_id' => 'integer|min:1|exists:users,id|required',
             'title'   => 'string|required',
             'body'    => 'string|required'
         ]);
     }
-
-
-
-
-
-
 
 }
