+++ app/Http/Controllers/UserController.php
@@ -3,6 +3,8 @@ namespace App\Http\Controllers;
 
 use App\Interfaces\UserRepositoryInterface;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller {
 
@@ -31,15 +33,28 @@ class UserController extends Controller {
 
     public function update(Request $request, $userId)
     {
-        $this->validator($request);
+        $user = $this->userRepository->find($userId);
 
-        return $this->userRepository->updateUser($request->toArray(), $userId) ?
+        if(Gate::denies('update', $user))
+            return response(['message' => 'Unauthorized'], 401);
+
+
+        $validatedRequest = $this->validate($request, [
+            'full_name' => 'string|required'
+        ]);
+
+        return $this->userRepository->updateUser($validatedRequest, $userId) ?
             response(['message' => 'User updated successfully'], 201):
             response(['message' => 'User not found'], 404);
     }
 
     public function destroy($userId)
     {
+        $user = $this->userRepository->find($userId);
+
+        if(Gate::denies('delete', $user))
+            return response(['message' => 'Unauthorized'], 401);
+
         return $this->userRepository->deleteUser($userId) ?
             response(['message' => 'User deleted successfully'], 201):
             response(['message' => 'User not found'], 404);
@@ -49,7 +64,7 @@ class UserController extends Controller {
     {
         return $this->validate($request, [
             'full_name' => 'string|required',
-            'email'     => 'email|required',
+            'email'     => 'email|required|unique:users,email',
             'password'  => 'string|required',
         ]);
